server/requirements.txt
Flask
opencv-python
pose-format
tqdm
mediapipe
openai
python-dotenv
azure-cognitiveservices-speech
moviepy